requirements.txt
requests==2.32.3
pymysql==1.1.1
python-dotenv==1.0.1
